Build(Core): Make packages into Monorepo
The core package should be responsible only for basic functionality: i.e. determining whether a color is compliant or not. For the visual comparison, it makes sense to create a separate application entirely and pull in core as a dependency. To manage this dependency tree, I am opting to use Lerna to create a Monorepo style repository.